# Improve speed of native WebSocket frame parsing