## What do these changes do?

Improve speed for native parsing WebSocket frames using faster bit-manipulation and constants. Doesn't require any new documentation/unit-tests.

## Are there changes in behavior for the user?

No changes in behavior.

## Related issue number

No issues are closed with this PR.

## Checklist

- [x] I think the code is well written
- [ ] Unit tests for the changes exist (Not needed?)
- [ ] Documentation reflects the changes (Not needed?)
- [x] Add yourself to `CONTRIBUTORS.txt`
- [x] Add a new entry to `CHANGES.rst`