@asvetlov 2016-09-14
1. I like using hex numbers for bits, masks etc. Why did you replace them with decimals?
2. What is speed improvement? Do you have any benchmark?