@SethMichaelLarson 2016-09-14
@asvetlov I will change the decimal back to hex. 
The other changes are just using bit arithmetic instead of integer arithmetic. Simple timeit tests can show that this is substantially faster.