@asvetlov 2016-09-14
If the difference is huge -- please provide a benchmark for the function.
I doubt that integer arithmetic is slower that bit based on current CPU.