@SethMichaelLarson 2016-09-14
The test that failed in Travis is an asyncio timeout test, not anything related to WebSockets.
You're right that the difference isn't huge, but free cycles are still free cycles.

Tested on a Intel Xeon X5650 6 x 2.67GHz
Before:
```
$ python3.5 -m timeit -s 's = 30' 's // 4'
10000000 loops, best of 3: 0.0644 usec per loop

$ python3.5 -m timeit -s 's = 3289475' 's % 4'
10000000 loops, best of 3: 0.0662 usec per loop
```

After:
```
$ python3.5 -m timeit -s 's = 30' 's >> 2'
10000000 loops, best of 3: 0.0549 usec per loop

$ python3.5 -m timeit -s 's = 3289475' 's & 3'
10000000 loops, best of 3: 0.0586 usec per loop
```

17% speed-up for // vs. >> and 12% speed-up for % vs. &