@asvetlov 2016-09-14
I mean the benchmark for the whole `parse_frame`, not single math operation only.
Maybe rewriting with cython will make a value but your change is just invisible.